+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/OspfExternalRoute.java
@@ -223,7 +223,7 @@ public abstract class OspfExternalRoute extends OspfRoute {
   }
 
   @Override
-  public int getTag() {
+  public long getTag() {
     return NO_TAG;
   }
 }
